revert back to non split entities
